npm & unpkg

+++ scripts/serve.sh
@@ -1,18 +1,30 @@
 #!/bin/bash
 
 # args
-port=${1:-1234}
-dist=${2:-"./dist"}
-html=${3:-"test.html"}
+port=$1
+html=$2
+dist=$3
 
 # config
 bin=live-server
-watch=$dist,$html
+
+# functions
+error() {
+    echo "Error, $1"
+    exit 1
+}
+
+# check
+[[ -n "$port" ]] || error "Port number required!"
+[[ -n "$html" ]] || error "HTML path required!"
 
 # existing
 kill `pgrep -f $bin` 2>/dev/null
 
 # main
+watch=`[[ -n "$dist" ]] && echo $html,$dist || echo $html`
+echo $html
+echo $watch
 npx $bin --port=$port --watch=$watch --no-browser --entry-file=$html &
 pid=$!
 sleep 0.5
